@@ -1,14 +1,14 @@
 var webpack = require('webpack');
 var config = require('./base');
+var path = require("path");
 var CopyWebpackPlugin = require('copy-webpack-plugin');
-var ExtractTextPlugin = require('extract-text-webpack-plugin');
+var ExtractTextPlugin = require('extract-text-webpack-plugin');   //单独分离css
 
 /**
  * ouput config
  * @type {String}
  */
-config.output.filename = '[name].[chunkhash:6].js';
-config.output.chunkFilename = '[id].[chunkhash:6].js';
+ config.output.filename = '[name].[chunkhash:6].js';
 
 /**
  * devtool config
@@ -20,6 +20,7 @@ config.devtool = "cheap-source-map";
  * loaders config
  * @type {RegExp}
  */
+
 config.module.loaders.push({
   test: /\.css$/,
   loader: ExtractTextPlugin.extract('style', 'css')
@@ -27,6 +28,8 @@ config.module.loaders.push({
   test: /\.less$/,
   loader: ExtractTextPlugin.extract('style', 'css!less')
 });
+
+
 
 /**
  * plugins config
@@ -40,18 +43,18 @@ config.plugins.push(
       ignore: ['*.md']
     }
   ]),
-  new webpack.optimize.DedupePlugin(),
-  new webpack.optimize.UglifyJsPlugin({
+  new webpack.optimize.DedupePlugin(),    //去重
+  new webpack.optimize.UglifyJsPlugin({   //压缩
     compress: {
       warnings: false
     }
   }),
-  new webpack.optimize.OccurenceOrderPlugin(),
-  // 公共代码分离打包
-  new webpack.optimize.CommonsChunkPlugin({
-    names: ['vendor', 'mainifest']
-  }),
-  // 若要按需加载 CSS 则请注释掉该行
+  new webpack.optimize.OccurenceOrderPlugin(),    //分配最小id
+  //公共代码分离打包
+  // new webpack.optimize.CommonsChunkPlugin({
+  //   names: ['vendor', 'mainifest']
+  // }),
+  //若要按需加载 CSS 则请注释掉该行
   new ExtractTextPlugin('[name].[contenthash:6].css', {
     allChunks : true
   })
